requirements.txt
paho-mqtt
adafruit-circuitpython-ads1x15
adafruit-circuitpython-servokit
adafruit-circuitpython-neopixel
psutil
gpiozero
pyfiglet
watchdog